trying a lot of stuff out / now working on calculations
As mentioned in a short message. we need server access to install nodeJS on the server and run it over node in order to make the whole thing work.
as we do not have server access, we can't do that now.
I'm working on the calculations right now, the totalTIme can already be calculated