working on calculations
Now we are able to calculate the prices for the machines according to teh minutes used. Therefore I use a percentage calculation.
I take the minutes written in the textfield and divide through 1 hour or rather 60mins. This makes it easier to calculate the price for the machine
as the price for 1 machine over a timespan of 1 hour costs 5€. That means if the user inputs 60 in the minutes textfield. We have a percentage of 1%.
then we simply multiply the (price for one machine per hour * percentage * quantity)

for the workers we nearly use the same approach